Add files via upload
Write a one-file C or C++ program. Name it (first_name)_(last_name)_longest.c. Your program should print out the 10 longest words from the words file.  Words can be printed in any order, and if there are multiple words tied for 10th longest print any of those that are tied for that length.  Do not print any other output. The program reads the name of the words file from the command line, and print results to stdout.